clean up account type checking functions
